Moved attached property to RaygunClient and logic to RUMService
