+++ Raygun4UWP/Raygun4UWP.csproj
@@ -146,7 +146,6 @@
     <Compile Include="RUM\PayloadModels\RaygunRUMMessage.cs" />
     <Compile Include="RUM\PayloadModels\RaygunRUMTimingData.cs" />
     <Compile Include="RUM\PayloadModels\RaygunRUMTimingInfo.cs" />
-    <Compile Include="RUM\RaygunRUM.cs" />
     <Compile Include="RUM\RUMService.cs" />
     <EmbeddedResource Include="Properties\Raygun4UWP.rd.xml" />
   </ItemGroup>
